Separate interface and manager modules
